Use git version of django-mailman3
The feature in the released Hyperkitty to disable gravatar depends on a
commit in django_mailman3 which is not yet released. If that commit is
not used then an error will be thrown if gravatar is disabled but the
user tries to navigate to the account page. So use an intermediate
version that includes the missing commit and name it -r1.